backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-genai==0.2.2
pillow==10.1.0
slowapi==0.1.9
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-fastapi-instrumentator==6.1.0
structlog==23.2.0

